[FAB2023-999] fix : conflict 해결
