+++ project/ovp/server/src/main/java/com/mobigen/ovp/common/openmete_client/ClassificationClient.java
@@ -1,19 +1,21 @@
 package com.mobigen.ovp.common.openmete_client;
 
 import com.mobigen.ovp.common.openmete_client.dto.classification.ClassificationData;
+import com.mobigen.ovp.common.openmete_client.dto.classification.ClassificationEdit;
 import com.mobigen.ovp.common.openmete_client.dto.classification.detail.ClassificationDetail;
 import com.mobigen.ovp.common.openmete_client.dto.classification.detail.ClassificationDetailByName;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
+import java.util.Map;
 
 @FeignClient(name = "ClassificationClient", url = "${properties.ovp.open-metadata-url}/classifications")
 public interface ClassificationClient {
@@ -63,7 +65,4 @@ public interface ClassificationClient {
      */
     @DeleteMapping("/{id}")
     ResponseEntity<Void> deleteClassification(@PathVariable String id, @RequestParam(defaultValue = "true") boolean recursive, @RequestParam(defaultValue = "true") boolean hardDelete);
-
-    //    @PostMapping("")
-//    Object createTag(@RequestBody Map<String, Object> params);
 }
