+++ project/ovp/server/src/main/java/com/mobigen/ovp/common/openmete_client/TagClient.java
@@ -6,6 +6,8 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Map;
@@ -27,7 +29,8 @@ public interface TagClient {
      * @return
      */
     @GetMapping("")
-    ClassificationTag getClassificationTags(
-            @RequestParam(defaultValue = "da") String parent);
+    ClassificationTag getClassificationTags(@RequestParam(defaultValue = "da") String parent);
 
+    @PostMapping("")
+    Object createTag(@RequestBody Map<String, Object> params);
 }
